server/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
websockets>=12.0
pyjwt>=2.8
passlib[bcrypt]>=1.7
